ai/requirements.txt
python-dotenv==1.1.1
uagents==0.22.7
